+++ hedera-mirror-test/src/test/java/com/hedera/mirror/test/e2e/acceptance/steps/BaseContractFeature.java
@@ -23,7 +23,7 @@ import com.hedera.mirror.rest.model.ContractResult;
 import com.hedera.mirror.test.e2e.acceptance.util.FeatureInputHandler;
 import java.util.List;
 
-public abstract class BaseContractFeature extends AbstractFeature {
+public abstract class BaseContractFeature extends AbstractEstimateFeature {
     protected DeployedContract deployedParentContract;
 
     protected ContractResponse verifyContractFromMirror(boolean isDeleted) {
