+++ hedera-mirror-test/src/test/java/com/hedera/mirror/test/e2e/acceptance/steps/EstimateFeature.java
@@ -800,7 +800,7 @@ public class EstimateFeature extends AbstractEstimateFeature {
         IERC20_TOKEN_ASSOCIATE("associate()", 728033),
         IERC20_TOKEN_DISSOCIATE("dissociate()", 728033),
         CREATE_CHILD("createChild", 0),
-        GET_BYE_CODE("getBytecode", 0),
+        GET_BYTE_CODE("getBytecode", 0),
         GET_ADDRESS("getAddress", 0),
         CREATE_2_DEPLOY("create2Deploy", 0),
         VACATE_ADDRESS("vacateAddress", 0);
